+++ .storybook/components/Markdown/Markdown.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import ReactMarkdown from 'react-markdown'
+import remarkGfm from 'remark-gfm'
 import { BaseProps } from '@toptal/picasso-shared'
 import { makeStyles, Theme } from '@material-ui/core/styles'
 
@@ -17,7 +18,11 @@ const Markdown = (props: Props): JSX.Element => {
 
   const classes = useStyles()
 
-  return <ReactMarkdown className={classes.root}>{children}</ReactMarkdown>
+  return (
+    <ReactMarkdown className={classes.root} remarkPlugins={[remarkGfm]}>
+      {children}
+    </ReactMarkdown>
+  )
 }
 
 Markdown.displayName = 'Markdown'